requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
scipy
scikit-learn
openpyxl
pyyaml
tabulate